requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
bcrypt
python-jose
python-dotenv
psycopg2-binary
